Add working SSH server that lists keys
Flesh out a working SSH server that lists the fingerprints and key
length of RSA, DSA or ECDSA keys presented to it.

Also, warn if the client has enabled agent forwarding or X11 forwarding
using techniques borrowed from FiloSottile/whosthere.

Output so far looks something like:

    $ ssh localhost:2022
    The public keys presented by your SSH client are:

    Bits  Type     Fingerprint
    4096  ssh-rsa  ed:9a:d2:5d:7b:c0:e5:cf:b9:bc:5c:6b:ce:3a:db:20
    1024  ssh-rsa  ad:6a:64:d0:b7:8a:73:b4:8a:d6:c6:ff:50:f3:4e:cd
    Connection to localhost closed.

I'm not a cryptographer nor an SSH expert but I believe the code to
determine the public key lengths is reasonably reliable according to my
testing:

For RSA, use the bit length of the modulus `n` [1][2], where the public
key consists of `n` and the public exponent `e`.

For DSA, use `p` [3][4], where the public key consists of `p`, `q` and
`g` [3][4].

For ECDSA, we can use `k.Params().BitSize` [5].

[1]: https://en.wikipedia.org/wiki/RSA_(cryptosystem)#Key_generation
[2]: https://github.com/golang/crypto/blob/83f1503f771a82af8a31f358eb825e9efb5dae6c/openpgp/packet/public_key.go#L700
[3]: https://github.com/golang/crypto/blob/83f1503f771a82af8a31f358eb825e9efb5dae6c/openpgp/packet/public_key.go#L702
[4]: https://www.dlitz.net/software/pycrypto/api/current/Crypto.PublicKey.DSA-module.html
[5]: https://github.com/golang/crypto/blob/83f1503f771a82af8a31f358eb825e9efb5dae6c/ssh/keys.go#L395-L402