Print warnings for RSA keys with 2048 bits or less
RSA keys with fewer than 2048 bits are generally considered weak by
today's standards [1]. The default for OpenSSH when generating RSA keys
is 2048 bits, so warn the user if their keys have lengths shorter than
that.

This adds an 'Issues' column in the output so that we can indicate which
keys are weak.

[1]: https://www.schneier.com/blog/archives/2013/09/the_nsas_crypto_1.html