Return immediately if handshake fails
Otherwise, the server effectively hangs and stops serving new requests
because the `sessions.mu` lock becomes deadlocked.

It was possible to trigger this behaviour by connecting to the server
then closing the connection before the handshake had completed, which
shows up as `EOF` in the error returned by `ssh.NewServerConn()`.