Use a map to make blacklisting faster
The operation to iteratively search through the blacklisted keys and
compare each one to the public keys presented by the user is noticeably
slow.

Make it faster by using a map of booleans indexed by the blacklisted key
as a string. If a given key is not in the blacklist, `blacklist[key]`
will evaluate to false.

This should be an O(1) operation rather than a O(n) operation as it was
previously.